requirements.txt
rpi-lgpio>=0.6
Flask>=2.0.0